backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
jinja2==3.1.3
python-multipart==0.0.9
python-dotenv==1.0.1
